:fire: merge develop into main 4

+++ src/main/java/com/mallang/mallang_backend/domain/sentence/expressionbook/service/impl/ExpressionBookServiceImpl.java
@@ -2,12 +2,12 @@ package com.mallang.mallang_backend.domain.sentence.expressionbook.service.impl;
 
 import com.mallang.mallang_backend.domain.member.entity.Member;
 import com.mallang.mallang_backend.domain.member.repository.MemberRepository;
+import com.mallang.mallang_backend.domain.sentence.expression.entity.Expression;
 import com.mallang.mallang_backend.domain.sentence.expression.repository.ExpressionRepository;
 import com.mallang.mallang_backend.domain.sentence.expressionbook.dto.ExpressionBookRequest;
 import com.mallang.mallang_backend.domain.sentence.expressionbook.dto.ExpressionBookResponse;
 import com.mallang.mallang_backend.domain.sentence.expressionbook.dto.ExpressionResponse;
 import com.mallang.mallang_backend.domain.sentence.expressionbook.dto.savedExpressionsRequest;
-import com.mallang.mallang_backend.domain.sentence.expression.entity.Expression;
 import com.mallang.mallang_backend.domain.sentence.expressionbook.entity.ExpressionBook;
 import com.mallang.mallang_backend.domain.sentence.expressionbook.repository.ExpressionBookRepository;
 import com.mallang.mallang_backend.domain.sentence.expressionbook.service.ExpressionBookService;
@@ -19,13 +19,15 @@ import com.mallang.mallang_backend.domain.video.video.repository.VideoRepository
 import com.mallang.mallang_backend.global.exception.ErrorCode;
 import com.mallang.mallang_backend.global.exception.ServiceException;
 import com.mallang.mallang_backend.global.gpt.service.GptService;
-
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalTime;
 import java.util.List;
+
+import static com.mallang.mallang_backend.global.constants.AppConstants.DEFAULT_EXPRESSION_BOOK_NAME;
+import static com.mallang.mallang_backend.global.exception.ErrorCode.EXPRESSIONBOOK_DELETE_DEFAULT_FORBIDDEN;
 
 @Service
 @RequiredArgsConstructor
@@ -43,7 +45,19 @@ public class ExpressionBookServiceImpl implements ExpressionBookService {
     @Transactional
     public ExpressionBookResponse create(ExpressionBookRequest request, Long memberId) {
         Member member = memberRepository.findById(memberId)
-                .orElseThrow(() -> new ServiceException(ErrorCode.USER_NOT_FOUND));
+                .orElseThrow(() -> new ServiceException(ErrorCode.MEMBER_NOT_FOUND));
+
+        if (!member.getSubscription().isStandardOrHigher()) {
+            throw new ServiceException(ErrorCode.NO_EXPRESSIONBOOK_CREATE_PERMISSION);
+        }
+
+        if (request.getName().equals(DEFAULT_EXPRESSION_BOOK_NAME)) {
+            throw new ServiceException(ErrorCode.EXPRESSIONBOOK_CREATE_DEFAULT_FORBIDDEN);
+        }
+
+        if (expressionBookRepository.existsByMemberAndName(member, request.getName())) {
+            throw new ServiceException(ErrorCode.DUPLICATE_EXPRESSIONBOOK_NAME);
+        }
 
         ExpressionBook expressionBook = ExpressionBook.builder()
                 .name(request.getName())
@@ -60,7 +74,7 @@ public class ExpressionBookServiceImpl implements ExpressionBookService {
     @Transactional
     public List<ExpressionBookResponse> getByMember(Long memberId) {
         Member member = memberRepository.findById(memberId)
-                .orElseThrow(() -> new ServiceException(ErrorCode.USER_NOT_FOUND));
+                .orElseThrow(() -> new ServiceException(ErrorCode.MEMBER_NOT_FOUND));
 
         List<ExpressionBook> books = expressionBookRepository.findAllByMember(member);
 
@@ -91,6 +105,12 @@ public class ExpressionBookServiceImpl implements ExpressionBookService {
         if (!book.getMember().getId().equals(memberId)) {
             throw new ServiceException(ErrorCode.FORBIDDEN_EXPRESSION_BOOK);
         }
+
+        if (DEFAULT_EXPRESSION_BOOK_NAME.equals(book.getName())) {
+            throw new ServiceException(EXPRESSIONBOOK_DELETE_DEFAULT_FORBIDDEN);
+        }
+
+        expressionBookItemRepository.deleteAllById_ExpressionBookId(expressionBookId);
 
         expressionBookRepository.delete(book);
     }
